Fix session expiry bug.
AJAX call that was supposed to cause a timeout was actually resetting
the timer.  Change calls so that the keep alive call attached to all
AJAX requests does not do anything if a warning is currently being
shown.  At that point an explicit user action is required to keep
session going.