SNO-385: Rewrite session expiry warning strategy to handle multiple windows.
Last activity time is maintained in the CwmSession; it is updated by
most events.  AJAX activity that does not log an event but should keep
the session active should call ICwmSessionService.registerActivity().
SessionExpireWarningDialog will always check with the server so that
a background window cannot decide that you are inactive and kill the
session while you're working in another window.